+++ packages/gallery/tests/e2e/styleParams/imageInfoType.e2e.spec.js
@@ -22,7 +22,6 @@ describe('imageInfoType - e2e', () => {
       itemBorderWidth: 5,
       textBoxBorderWidth: 5,
       titlePlacement: GALLERY_CONSTS.placements.SHOW_BELOW,
-      oneRow: false,
       scrollDirection: GALLERY_CONSTS.scrollDirection.VERTICAL,
     });
     await driver.waitFor.hookToBeVisible('item-container');
@@ -37,7 +36,6 @@ describe('imageInfoType - e2e', () => {
       itemBorderWidth: 5,
       textBoxBorderWidth: 5,
       titlePlacement: GALLERY_CONSTS.placements.SHOW_BELOW,
-      oneRow: false,
       scrollDirection: GALLERY_CONSTS.scrollDirection.VERTICAL,
     });
     await driver.waitFor.hookToBeVisible('item-container');
@@ -52,7 +50,6 @@ describe('imageInfoType - e2e', () => {
       itemBorderWidth: 5,
       textBoxBorderWidth: 5,
       titlePlacement: GALLERY_CONSTS.placements.SHOW_BELOW,
-      oneRow: false,
       scrollDirection: GALLERY_CONSTS.scrollDirection.VERTICAL,
     });
     await driver.waitFor.hookToBeVisible('item-container');
@@ -67,7 +64,6 @@ describe('imageInfoType - e2e', () => {
       itemBorderWidth: 5,
       textBoxBorderWidth: 5,
       titlePlacement: GALLERY_CONSTS.placements.SHOW_BELOW,
-      oneRow: false,
       scrollDirection: GALLERY_CONSTS.scrollDirection.VERTICAL,
     });
     await driver.waitFor.hookToBeVisible('item-container');