+++ packages/gallery/tests/styleParams/scrollDirection.spec.js
@@ -2,8 +2,8 @@ import GalleryDriver from '../drivers/reactDriver';
 import { expect } from 'chai';
 import { images2 } from '../drivers/mocks/items';
 import { styleParams, container } from '../drivers/mocks/styles';
-
-describe('styleParam - oneRow', () => {
+import { GALLERY_CONSTS } from 'pro-gallery-lib';
+describe('styleParam - scrollDirection', () => {
   let driver;
   const initialProps = {
     container,
@@ -15,10 +15,9 @@ describe('styleParam - oneRow', () => {
     driver = new GalleryDriver();
   });
 
-  it('should render element "#gallery-horizontal-scroll" when "oneRow" is "true"', async () => {
+  it('should render element "#gallery-horizontal-scroll" when "scrollDirection" is horizontal', async () => {
     Object.assign(initialProps.styles, {
-      oneRow: true,
-      scrollDirection: null, // make sure scroll direction doesnt change oneRow
+      scrollDirection: GALLERY_CONSTS.scrollDirection.HORIZONTAL,
       galleryLayout: 2,
     });
     driver.mount.proGallery(initialProps);
@@ -27,10 +26,9 @@ describe('styleParam - oneRow', () => {
     expect(elem).to.have.lengthOf(1);
     driver.detach.proGallery();
   });
-  it('should render element "#pro-gallery-margin-container" when "oneRow" is "false"', async () => {
+  it('should render element "#pro-gallery-margin-container" when "scrollDirection" is vertical', async () => {
     Object.assign(initialProps.styles, {
-      oneRow: false,
-      scrollDirection: null, // make sure scroll direction doesnt change oneRow
+      scrollDirection: GALLERY_CONSTS.scrollDirection.VERTICAL,
       galleryLayout: 2,
     });
     driver.mount.proGallery(initialProps);
@@ -39,10 +37,9 @@ describe('styleParam - oneRow', () => {
     expect(elem).to.have.lengthOf(1);
     driver.detach.proGallery();
   });
-  it('should set margin on items to "0px" when oneRow is false(vertical)', async () => {
+  it('should set margin on items to "0px" when scrollDirection is false(vertical)', async () => {
     Object.assign(initialProps.styles, {
-      oneRow: false,
-      scrollDirection: null, // make sure scroll direction doesnt change oneRow
+      scrollDirection: GALLERY_CONSTS.scrollDirection.VERTICAL,
       galleryLayout: 2,
     });
     driver.mount.proGallery(initialProps);