+++ packages/layouts/src/logic/layouter.js
@@ -5,6 +5,7 @@ import { Group } from './group.js';
 import { Strip } from './strip.js';
 import { Column } from './column.js';
 import layoutsStore from './layoutsStore.js';
+import { GALLERY_CONSTS } from 'pro-gallery-lib';
 
 export default class Layouter {
   constructor(layoutParams) {
@@ -134,7 +135,7 @@ export default class Layouter {
           });
         } else {
           //---------------------| STRIPS GALLERY |----------------------//
-          if (this.styleParams.oneRow) {
+          if (this.scrollDirection === GALLERY_CONSTS.scrollDirection.HORIZONTAL) {
             //remove items from the last group:
             const lastGroups = this.groups.slice(-1);
             lastGroups.forEach((group) => {
@@ -186,7 +187,7 @@ export default class Layouter {
                 idx: this.strips.length + 1,
                 container: this.container,
                 groupsPerStrip: this.styleParams.groupsPerStrip,
-                oneRow: this.styleParams.oneRow,
+                scrollDirection: this.styleParams.scrollDirection,
                 targetItemSize: this.targetItemSize,
               });
             }
@@ -241,7 +242,7 @@ export default class Layouter {
         idx: 1,
         container: this.container,
         groupsPerStrip: this.styleParams.groupsPerStrip,
-        oneRow: this.styleParams.oneRow,
+        scrollDirection: this.styleParams.scrollDirection,
         targetItemSize: this.targetItemSize,
       });
 
@@ -382,7 +383,7 @@ export default class Layouter {
             idx: this.strip.idx + 1,
             container: this.container,
             groupsPerStrip: this.styleParams.groupsPerStrip,
-            oneRow: this.styleParams.oneRow,
+            scrollDirection: this.styleParams.scrollDirection,
             targetItemSize: this.targetItemSize,
           });
 
@@ -401,7 +402,7 @@ export default class Layouter {
         this.strip.addGroup(this.group);
 
         if (this.isLastImage && this.strip.hasGroups) {
-          if (this.styleParams.oneRow) {
+          if (this.styleParams.scrollDirection === GALLERY_CONSTS.scrollDirection.HORIZONTAL) {
             this.strip.height =
               this.container.galleryHeight +
               (this.styleParams.imageMargin / 2 -