+++ packages/lib/src/settings/options/index.js
@@ -5,7 +5,6 @@ import enableInfiniteScroll from './enableInfiniteScroll';
 import cubeType from './cubeType';
 import cubeImages from './cubeImages';
 import cubeRatio from './cubeRatio';
-import oneRow from './oneRow';
 import titlePlacement from './titlePlacement';
 import galleryMargin from './galleryMargin';
 import isVertical from './isVertical';
@@ -89,6 +88,7 @@ import pauseAutoSlideshowOnHover from './pauseAutoSlideshowOnHover';
 import showVideoControls from './showVideoControls';
 import shouldIndexDirectShareLinkInSEO from './shouldIndexDirectShareLinkInSEO';
 import slideTransition from './slideTransition';
+import useMaxDimensions from './useMaxDimensions';
 
 export default {
   itemClick,
@@ -98,7 +98,6 @@ export default {
   cubeType,
   cubeImages,
   cubeRatio,
-  oneRow,
   titlePlacement,
   galleryMargin,
   isVertical,
@@ -182,6 +181,7 @@ export default {
   pauseAutoSlideshowOnHover,
   shouldIndexDirectShareLinkInSEO,
   slideTransition,
+  useMaxDimensions,
 };
 
 // TODO = add the options: