+++ packages/lib/src/settings/options/numberOfImagesPerRow.js
@@ -1,11 +1,12 @@
 import { INPUT_TYPES } from '../utils/constants';
+import { default as GALLERY_CONSTS } from '../../common/constants';
 
 export default {
   title: 'Images Per Row',
   isRelevantDescription:
     'Set a Vertical gallery ("Scroll Direction" as "Vertical"), set "Layout Orientation" to "Column" and set "Max Group Size" to be "1".',
   isRelevant: (styleParams) =>
-    !styleParams.oneRow &&
+    styleParams.scrollDirection === GALLERY_CONSTS.scrollDirection.VERTICAL &&
     styleParams.isVertical &&
     styleParams.gridStyle === 1,
   type: INPUT_TYPES.NUMBER,