+++ packages/playground/src/components/App/App.js
@@ -308,6 +308,7 @@ export function App() {
 }
 
 const addResizable = (Component, props, resizedDims, setResizedDims, gallerySettings) => {
+  props.shouldValidateTypes = false
   return gallerySettings.responsivePreview ? (<div style={{
     background: '#666',
     width: '100%',