+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+## [v3.1.38](https://pro-gallery-3-1-38.surge.sh) (07/07/2021)
+ 
+#### GALLERY
+ -  fix: use imageDimensions in video item and text item
+
+---
+## [v3.1.37](https://pro-gallery-3-1-37.surge.sh) (07/07/2021)
+ 
+#### GALLERY
+ -  fix: Navigation Arrows Missing When Navigating to Page with Slider Gallery
+
+#### OTHER
+ -  feat:added alt text 
+
+---
+## [v3.1.36](https://pro-gallery-3-1-36.surge.sh) (06/07/2021)
+ 
+#### OTHER
+ -  improve: refactor image dimensions to layouter and added useMaxDImensions styleParam
+
+---
+## [v3.1.35](https://pro-gallery-3-1-35.surge.sh) (04/07/2021)
+ 
+#### GALLERY
+ - fix: destruct arrowsColor from styleParams and not props
+
+---
+## [v3.1.34](https://pro-gallery-3-1-34.surge.sh) (04/07/2021)
+ 
+#### OTHER
+ -  slideshowView: arrows -> fix container size to be same as svg size (to support arrowsPadding correctly) 
+
+---
+## [v3.1.33](https://pro-gallery-3-1-33.surge.sh) (04/07/2021)
+ 
+#### GALLERY
+ - fix: fix behaviour of what happens on the first tap when hovering behaviour set to NO_CHANGE
+
+---
+## [v3.1.32](https://pro-gallery-4-0-5.surge.sh) (01/07/2021)
+ 
+#### GALLERY
+ -  fix: Increasing removeArrowsIfNeeded setTimeOut call
+
+---
 ## [v3.1.31](https://pro-gallery-3-1-31.surge.sh) (30/06/2021)
 
 #### GALLERY